refactor: move funding UI into feature module
