+++ src/features/funding/components/FundingProjectDetail/ProjectTabs.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '../ui/tabs';
-import { Card, CardContent, CardHeader, CardTitle } from '../ui/card';
-import { Badge } from '../ui/badge';
-import { Avatar, AvatarFallback } from '../ui/avatar';
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { Badge } from '@/components/ui/badge';
+import { Avatar, AvatarFallback } from '@/components/ui/avatar';
 import { FileText, DollarSign, CheckCircle, TrendingUp } from 'lucide-react';
-import { FundingProject } from '../../types/fundingProject';
+import { FundingProject } from '@/types/fundingProject';
 
 interface ProjectTabsProps {
     project: FundingProject;
