+++ src/features/funding/components/FundingProjectDetail/index.ts
@@ -1,0 +1,3 @@
+export { FundingProjectDetail } from './FundingProjectDetail';
+export { ProjectHeader } from './ProjectHeader';
+export { ProjectTabs } from './ProjectTabs';
