+++ src/features/funding/components/PaymentModal/index.ts
@@ -1,0 +1,1 @@
+export { PaymentModal } from './PaymentModal';
